+++ AsBuiltReport.Microsoft.Azure/Src/Private/Report/Get-AbrAzSAQueue.ps1
@@ -36,7 +36,8 @@ function Get-AbrAzSAQueue {
 
     process {
         Try {
-            $AzSAQueues = Get-AzStorageAccount -ResourceGroupName $ResourceGroupName -StorageAccountName $StorageAccountName | Get-AzStorageQueue -ErrorAction SilentlyContinue | Sort-Object Name
+            $AzStorageContext = New-AzStorageContext -StorageAccountName $StorageAccountName -UseConnectedAccount -ErrorAction SilentlyContinue
+            $AzSAQueues = Get-AzStorageQueue -Context $AzStorageContext -ErrorAction SilentlyContinue | Sort-Object Name
             if ($AzSAQueues) {
                 Write-PscriboMessage $LocalizedData.Collecting
                 Section -Style NOTOCHeading6 -ExcludeFromTOC $LocalizedData.Heading {